Implement the basic server framework
This does the work necessary to start up the internal Readium web
server, and sets up a web view to open what should be the initial URL
in the EPUB spine. Unfortunately, this currently results in an error
from the internal server as the URL is evidently not the correct one.